feat: allow 15 and 13 digits nir
